requirements.txt
numpy>=1.23,<2.0
scipy==1.11.4
pandas==1.5.3

scikit-learn==1.3.2
matplotlib==3.7.3
seaborn>=0.12.1,<0.13.0

torch
pysr
ipykernel
pyartemis

